feat: add tiptap editor
