+++ package.json
@@ -14,5 +14,10 @@
     "@tailwindcss/typography": "^0.5.9",
     "@types/node": "^18.17.3",
     "nuxt": "^3.6.5"
+  },
+  "dependencies": {
+    "@tiptap/pm": "^2.1.7",
+    "@tiptap/starter-kit": "^2.1.7",
+    "@tiptap/vue-3": "^2.1.7"
   }
 }
