+++ src/components/Editor/example.ts
@@ -1,0 +1,116 @@
+export const blogContent = `
+   <h1>Comparing Vue.js and React.js</h1>
+    <p>Posted on August 25, 2023</p>
+
+    <h2>Introduction to Vue.js and React.js</h2>
+    <p>
+      Vue.js and React.js are both popular JavaScript libraries used for
+      building user interfaces in web applications. They share many
+      similarities, but also have distinct differences that developers should
+      consider when choosing between them.
+    </p>
+
+    <h2>Similarities</h2>
+    <ul>
+      <li>
+        Component-Based Architecture: Both Vue.js and React.js encourage
+        building applications using reusable components, promoting modularity
+        and maintainability.
+      </li>
+      <li>
+        Virtual DOM: Both libraries utilize a virtual DOM to improve performance
+        by minimizing direct manipulation of the actual DOM.
+      </li>
+      <li>
+        Reactivity: Both Vue.js and React.js offer mechanisms to automatically
+        update the user interface when data changes, enhancing user experience.
+      </li>
+    </ul>
+
+    <h2>Differences</h2>
+    <h3>Template Syntax vs JSX</h3>
+    <p>
+      Vue.js uses a template syntax that closely resembles HTML, making it easy
+      to understand and transition for developers familiar with HTML. React.js,
+      on the other hand, uses JSX (JavaScript XML), which integrates HTML-like
+      syntax within JavaScript code.
+    </p>
+
+    <div>
+      <h4>Vue.js Template Example</h4>
+      <pre><code>
+&lt;template&gt;
+    &lt;div&gt;
+        &lt;p&gt;{{ message }}&lt;/p&gt;
+    &lt;/div&gt;
+&lt;/template&gt;
+
+&lt;script&gt;
+export default {
+    data() {
+        return {
+            message: 'Hello, Vue.js!'
+        };
+    }
+}
+&lt;/script&gt;
+        </code></pre>
+    </div>
+
+    <div>
+      <h4>React.js JSX Example</h4>
+      <pre><code>
+import React, { Component } from 'react';
+
+class App extends Component {
+    render() {
+        return (
+            &lt;div&gt;
+                &lt;p&gt;{this.state.message}&lt;/p&gt;
+            &lt;/div&gt;
+        );
+    }
+    
+    state = {
+        message: 'Hello, React.js!'
+    };
+}
+
+export default App;
+        </code></pre>
+    </div>
+
+    <h3>Learning Curve</h3>
+    <p>
+      Vue.js is often considered more approachable for beginners due to its
+      simple syntax and clear documentation. React.js might have a steeper
+      learning curve, especially for developers new to JSX and its functional
+      programming concepts.
+    </p>
+
+    <h3>Community and Ecosystem</h3>
+    <p>
+      React.js has a larger community and ecosystem, which translates to more
+      available libraries, tools, and resources. Vue.js, while growing rapidly,
+      has a comparatively smaller ecosystem.
+    </p>
+
+    <h3>Directives vs Props</h3>
+    <p>
+      Vue.js uses directives like <code>v-bind</code> and <code>v-on</code> to
+      handle data binding and event handling within templates. React.js uses
+      props to pass data from parent to child components, promoting a more
+      unidirectional data flow.
+    </p>
+
+    <h2>Conclusion</h2>
+    <p>
+      Both Vue.js and React.js are powerful tools for building modern web
+      applications. The choice between them depends on factors such as
+      familiarity with their syntax, project requirements, and personal
+      preferences. Developers should consider the similarities and differences
+      highlighted here to make an informed decision.
+    </p>
+
+    <p>Written by John Doe</p>
+`;
